==> src/MobiSegInsightsSE.etl.1.0/4-aux-gridded-population-processing.ipynb <==
# Process population data

%load_ext autoreload
%autoreload 2
%cd D:\mobi-social-segregation-se

# Load libs
import pandas as pd
import geopandas as gpd
import sqlalchemy
from lib import preprocess as preprocess

# Data location
user = preprocess.keys_manager['database']['user']
password = preprocess.keys_manager['database']['password']
port = preprocess.keys_manager['database']['port']
db_name = preprocess.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}?gssencmode=disable')

Population data from [Swedish University of Agricultural Sciences](https://maps.slu.se/).

pop_folder = "dbs/gridded_population/"
income_path = pop_folder + "income_2012_latest_shp/IH1_20191231_SweRef_shape/SLU_IH1_SweRef_2019_region.shp"
birth_path = pop_folder + "population_2013_latest_shp/B5RutaSW_20190101/B5_SweRef_region.shp"
job_path = pop_folder + "labour_market_2012_latest_shp/A4_20191231_SweRef_shape/A4_SweRef_2020_region.shp"

gdf_birth = gpd.read_file(birth_path)
gdf_job = gpd.read_file(job_path).rename(columns={'astRutstor': 'Rutstorl'})
gdf_income = gpd.read_file(income_path)

## 0. Deal with duplicated grids
### 0.1 Income file

gdf_income.loc[:, 'Totalt'] = gdf_income.loc[:, 'Kvartil_1'] + \
                              gdf_income.loc[:, 'Kvartil_2'] + \
                              gdf_income.loc[:, 'Kvartil_3'] + \
                              gdf_income.loc[:, 'Kvartil_4']

duplicated_grids = gdf_income.loc[gdf_income.duplicated(subset=['Ruta']), 'Ruta'].to_list()
# Non-duplicated grids
gdf_income_1 = gdf_income.loc[~gdf_income.Ruta.isin(duplicated_grids), :]
# Duplicated grids
gdf_income_2 = gdf_income.loc[gdf_income.Ruta.isin(duplicated_grids), :].sort_values(by=['Ruta'])

# Prepare a big grid data as a place holder
gdf_income_2_big = gdf_income_2.loc[gdf_income_2.Rutstorl == 1000, ['Ruta', 'Rutstorl', 'geometry']]

def merge_income_duplicated(data):
    if 0 in data.Totalt:
        return data[['Kvartil_1', 'Kvartil_2', 'Kvartil_3', 'Kvartil_4',
                     'Totalt', 'MedianInk', 'Tot_CDISP0']].sum()
    Kvartil_1 = data['Kvartil_1'].sum()
    Kvartil_2 = data['Kvartil_2'].sum()
    Kvartil_3 = data['Kvartil_3'].sum()
    Kvartil_4 = data['Kvartil_4'].sum()
    Totalt = data['Totalt'].sum()
    Tot_CDISP0 = data['Tot_CDISP0'].sum()
    MedianInk = sum(data['Totalt'] * data['MedianInk']) / data['Totalt'].sum()
    return pd.Series(dict(Kvartil_1=Kvartil_1,
                          Kvartil_2=Kvartil_2,
                          Kvartil_3=Kvartil_3,
                          Kvartil_4=Kvartil_4,
                          Totalt=Totalt,
                          MedianInk=MedianInk,
                          Tot_CDISP0=Tot_CDISP0))

gdf_income_2 = pd.merge(gdf_income_2_big,
                        gdf_income_2.groupby('Ruta').apply(merge_income_duplicated).reset_index(),
                        on='Ruta').dropna()

gdf_income = pd.concat([gdf_income_1, gdf_income_2])

### 0.2 Birth background file

gdf_birth.loc[:, 'TotFland'] = gdf_birth.loc[:, 'Sverige'] + \
                              gdf_birth.loc[:, 'Norden_uto'] + \
                              gdf_birth.loc[:, 'EU28_utom_'] + \
                              gdf_birth.loc[:, 'Övriga_vär']

duplicated_grids = gdf_birth.loc[gdf_birth.duplicated(subset=['Ruta']), 'Ruta'].to_list()
# Non-duplicated grids
gdf_birth_1 = gdf_birth.loc[~gdf_birth.Ruta.isin(duplicated_grids), :]
# Duplicated grids
gdf_birth_2 = gdf_birth.loc[gdf_birth.Ruta.isin(duplicated_grids), :].sort_values(by=['Ruta'])

# Prepare a big grid data as a place holder
gdf_birth_2_big = gdf_birth_2.loc[gdf_birth_2.Rutstorl == 1000, ['Ruta', 'Rutstorl', 'geometry']]

def merge_birth_duplicated(data):
    return data[['Sverige', 'Norden_uto', 'EU28_utom_', 'Övriga_vär', 'TotFland']].sum()

gdf_birth_2 = pd.merge(gdf_birth_2_big,
                        gdf_birth_2.groupby('Ruta').apply(merge_birth_duplicated).reset_index(),
                        on='Ruta').dropna()
gdf_birth = pd.concat([gdf_birth_1, gdf_birth_2])

### 0.3 Job file

gdf_job.loc[:, 'Totalt'] = gdf_job.loc[:, 'Offentliga'] + gdf_job.loc[:, 'Naringsliv']

duplicated_grids = gdf_job.loc[gdf_job.duplicated(subset=['Ruta']), 'Ruta'].to_list()
# Non-duplicated grids
gdf_job_1 = gdf_job.loc[~gdf_job.Ruta.isin(duplicated_grids), :]
# Duplicated grids
gdf_job_2 = gdf_job.loc[gdf_job.Ruta.isin(duplicated_grids), :].sort_values(by=['Ruta'])

# Prepare a big grid data as a place holder
gdf_job_2_big = gdf_job_2.loc[gdf_job_2.Rutstorl == 1000, ['Ruta', 'Rutstorl', 'geometry']]

def merge_job_duplicated(data):
    return data[['Offentliga', 'Naringsliv', 'Totalt']].sum()

gdf_job_2 = pd.merge(gdf_job_2_big,
                     gdf_job_2.groupby('Ruta').apply(merge_job_duplicated).reset_index(),
                     on='Ruta').dropna()
gdf_job = pd.concat([gdf_job_1, gdf_job_2])

### 0.4 Relationship between three gridding systems

They share a large number of grids.

shared_grids = dict(pop_in_job=len(gdf_birth.loc[gdf_birth.Ruta.isin(gdf_job.Ruta)]),
                    pop_in_income=len(gdf_birth.loc[gdf_birth.Ruta.isin(gdf_income.Ruta)]),
                    pop_total=len(gdf_birth))
shared_grids

The small grids of 250 m x 250 m account for 86.2% of Swedish population.

share_small_grids_in_pop = gdf_birth.loc[gdf_birth.Rutstorl == 250, 'TotFland'].sum() / (10.28*(10**6)) * 100
print("Share of small grids in total population: %.2f %%."%share_small_grids_in_pop)

share_nonzero_grids = dict(pop=len(gdf_birth.loc[gdf_birth.TotFland != 0]) / len(gdf_birth) * 100,
                           income=len(gdf_income.loc[gdf_income.Totalt != 0]) / len(gdf_income) * 100,
                           job=len(gdf_job.loc[gdf_job.Totalt != 0]) / len(gdf_job) * 100)
share_nonzero_grids

share_small_grids = dict(pop=len(gdf_birth.loc[gdf_birth.Rutstorl == 250]) / len(gdf_birth) * 100,
                         income=len(gdf_income.loc[gdf_income.Rutstorl == 250]) / len(gdf_income) * 100,
                         job=len(gdf_job.loc[gdf_job.Rutstorl == 250]) / len(gdf_job) * 100)
share_small_grids

## 1. Organize socio-economic information and the zoning system

gdf = gdf_income.copy()    #gdf_income.loc[gdf_income.Rutstorl == 250, :]
print("Number of grids: %s"%len(gdf))
gdf = pd.merge(gdf, gdf_birth.loc[:, ['Ruta', 'Sverige', 'Norden_uto', 'EU28_utom_', 'Övriga_vär', 'TotFland']],
               on='Ruta')
print("After adding birth background, number of grids: %s"%len(gdf))
gdf = pd.merge(gdf, gdf_job.loc[:, ['Ruta', 'Totalt']].rename(columns={'Totalt': 'Job_count'}),
               on='Ruta', how='left')
gdf.fillna(0, inplace=True)
gdf.iloc[0]

Socio-economic dimensions

cls = ['Kvartil_1', 'Kvartil_2', 'Kvartil_3', 'Kvartil_4',
       'Totalt', 'MedianInk', 'Tot_CDISP0',
       'Sverige', 'Norden_uto', 'EU28_utom_', 'Övriga_vär', 'TotFland', 'Job_count']

### 1.1 Find the overlapped small grids and big grids

gdf_overlapped = gpd.tools.overlay(gdf.loc[gdf['Rutstorl'] == 250,
                                           ['Ruta', 'geometry']].rename(columns={'Ruta': 'Ruta_s'}),
                                   gdf.loc[gdf['Rutstorl'] == 1000,
                                           ['Ruta', 'geometry']].rename(columns={'Ruta': 'Ruta_l'}), how='intersection')
df_overlapped = gdf_overlapped.dropna(how='any').drop(columns=['geometry'])

overlapped_ruta = df_overlapped.Ruta_s.to_list() + df_overlapped.Ruta_l.to_list()

### 1.2 Simplify the overlapped grids by merging them to big grids

df_overlapped_simp = pd.merge(df_overlapped,
                          gdf.loc[gdf['Rutstorl'] == 250, cls + ['Ruta']].rename(columns={'Ruta': 'Ruta_s'}),
                          on='Ruta_s')
df_overlapped_simp.columns

def merge_overlapped(data):
    if 0 in data.Totalt:
        return data[cls].sum()
    MedianInk = sum(data['Totalt'] * data['MedianInk']) / data['Totalt'].sum()
    return pd.Series(dict(Kvartil_1=data['Kvartil_1'].sum(),
                          Kvartil_2=data['Kvartil_2'].sum(),
                          Kvartil_3=data['Kvartil_3'].sum(),
                          Kvartil_4=data['Kvartil_4'].sum(),
                          Totalt=data['Totalt'].sum(),
                          MedianInk=MedianInk,
                          Tot_CDISP0=data['Tot_CDISP0'].sum(),
                          Sverige=data['Sverige'].sum(),
                          Norden_uto=data['Norden_uto'].sum(),
                          EU28_utom_=data['EU28_utom_'].sum(),
                          Övriga_vär=data['Övriga_vär'].sum(),
                          TotFland=data['TotFland'].sum(),
                          Job_count=data['Job_count'].sum()))

df_overlapped_simp = df_overlapped_simp.groupby('Ruta_l').apply(merge_overlapped).reset_index().rename(columns={'Ruta_l': 'Ruta'})
df_overlapped_simp.loc[:, 'Rutstorl'] = 1000
df_overlapped_simp.fillna(0, inplace=True)

gdf_no_overlap = gdf.loc[~gdf.Ruta.isin(overlapped_ruta), :]
df_overlap = gdf.loc[gdf.Ruta.isin(df_overlapped_simp.Ruta), cls + ['Ruta', 'Rutstorl']]
len(gdf_no_overlap), len(df_overlap)

# Merge overlapped small grids and big grids statistics
df_overlap = pd.concat([df_overlap, df_overlapped_simp])
df_overlap = df_overlap.groupby('Ruta').apply(merge_overlapped).reset_index()
df_overlap.loc[:, 'Rutstorl'] = 1000
df_overlap.fillna(0, inplace=True)

# Restore geometry
gdf_overlap = pd.merge(gdf.loc[gdf.Ruta.isin(df_overlap.Ruta), ['Ruta', 'geometry']],
                       df_overlap, on='Ruta')

# Create a new grid system with all ethnic and income statistics
gdf_final = pd.concat([gdf_no_overlap, gdf_overlap])
gdf_final.iloc[0]

### 1.3 Rename the columns

cols_selected = ['zone', 'grid_size', 'income_q1', 'income_q2', 'income_q3', 'income_q4', 'income_med',
                 'pop_income', 'birth_se', 'birth_nord', 'birth_eu', 'birth_other', 'pop', 'job', 'geometry']

gdf_final = gdf_final.rename(columns=dict(Ruta='zone',
                                          Rutstorl='grid_size',
                                          Kvartil_1='income_q1',
                                          Kvartil_2='income_q2',
                                          Kvartil_3='income_q3',
                                          Kvartil_4='income_q4',
                                          MedianInk='income_med',
                                          Totalt='pop_income',
                                          Sverige='birth_se',
                                          Norden_uto='birth_nord',
                                          EU28_utom_='birth_eu',
                                          Övriga_vär='birth_other',
                                          TotFland='pop',
                                          Job_count='job'))[cols_selected]
gdf_final.iloc[0]

### 1.4 Save the grids

# Save grids as .geojson and to database
gdf_final.to_postgis("grids", con=engine)
gdf_final.to_file('dbs/gridded_population/grids.geojson', driver='GeoJSON')
