==> src/MobiSegInsightsSE.etl.1.0/12-home-detection-and-filtering.ipynb <==
# Home detection
Temporal rules for home detection among top 3 visited clusters:

We infer the home area of each individual using its most common visited cluster between the hours of 10:00 p.m. and 6:00 a.m.

%load_ext autoreload
%autoreload 2
%cd D:\mobi-social-segregation-se

# Load libs
import pandas as pd
import numpy as np
import geopandas as gpd
import sqlalchemy
from tqdm import tqdm
from lib import preprocess as preprocess

## 1. Load temporal profiles

# Data location
user = preprocess.keys_manager['database']['user']
password = preprocess.keys_manager['database']['password']
port = preprocess.keys_manager['database']['port']
db_name = preprocess.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}?gssencmode=disable')

df = pd.read_sql_query(sql="""SELECT * FROM description.tempo_top3_p;""", con=engine)

df.head()

### 1.1 Summarise the weighted frequency count for the half-hour sequence 0-11 and 44-47

tempo_range = list(range(0, 12)) + list(range(44, 48))
def sum_home_stay(data):
    home_freq = data.loc[data.half_hour.isin(tempo_range), 'freq_wt'].sum()
    home_share = home_freq / data.freq_wt.sum() * 100
    return pd.Series(dict(home_freq=home_freq, home_share=home_share))
tqdm.pandas()
df_h = df.groupby(['uid', 'loc']).progress_apply(sum_home_stay).reset_index()

df_h.head()

## 2. Select home cluster and add zone information

def ind_select(data):
    home_ = data.loc[data.home_freq == data.home_freq.max(), ['loc', 'home_freq', 'home_share']].values[0]
    return pd.Series(dict(home=home_[0], home_freq=home_[1], home_share=home_[2]))
tqdm.pandas()
df_h_selected = df_h.groupby('uid').progress_apply(ind_select).reset_index().astype({"home": int})

df_h_selected.head()

### 2.1 Keep the top clusters with non-zero home-time stay

len(df_h_selected.loc[df_h_selected.home_freq > 0, :]) / len(df_h_selected)

df_h_selected = df_h_selected.loc[df_h_selected.home_freq > 0, :]

### 2.2 Add cluster characteristics

df_cls = pd.read_sql_query(sql="""SELECT * FROM description.clusters_top3_wt_p;""", con=engine)
df_cls.head()

df_home = pd.merge(df_h_selected, df_cls, left_on=['uid', 'home'], right_on=['uid', 'loc'])
df_home.head()

Get cluster centroids

df_stops = pd.read_sql_query(sql="""SELECT uid, lat, lng, loc FROM stops_p;""", con=engine)
df_stops = pd.merge(df_stops, df_home.loc[:, ['uid', 'loc']], on=['uid', 'loc'], how='inner')
df_stops.head()

df_stops = df_stops.drop_duplicates(subset=['uid', 'lat', 'lng', 'loc'])
df_stops.uid.nunique(), len(df_stops)

df_home = pd.merge(df_home, df_stops, on=['uid', 'loc'], how='left')

### 2.3 Add zone information - grid

gdf = gpd.GeoDataFrame.from_postgis("SELECT zone, geom FROM public.grids", con=engine).to_crs(4326)
gdf_home = preprocess.df2gdf_point(df_home, 'lng', 'lat', crs=4326, drop=True)
gdf_home = gpd.sjoin(gdf_home, gdf)
df_home_deso = pd.merge(df_home, gdf_home.loc[:, ['uid', 'zone']], on='uid', how='inner')

### 2.4 Add zone information - DeSO zone

gdf = gpd.GeoDataFrame.from_postgis("SELECT deso, geom FROM public.zones", con=engine).to_crs(4326)
gdf_home_ds = preprocess.df2gdf_point(df_home, 'lng', 'lat', crs=4326, drop=True)
gdf_home_ds = gpd.sjoin(gdf_home_ds, gdf)
df_home_deso = pd.merge(df_home_deso, gdf_home_ds.loc[:, ['uid', 'deso']], on='uid', how='inner')

### 2.5 Filtering

share_reliable_home = len(df_home_deso.loc[df_home_deso.freq >= 3, :]) / len(df_home_deso) * 100
print("Share of detected home locations with at least 3 appearances: %.2f %%"%share_reliable_home)

df_home_deso = df_home_deso.loc[df_home_deso.freq >= 3, :]
len(df_home_deso)

## 3. Individual weight
The inverse ratio of device number over population has extreme values in some regions, where only a few devices are included.
Weight trimming technique is applied:
Van de Kerckhove, Wendy, Leyla Mohadjer, and Thomas Krenzke. "A Weight Trimming Approach to Achieve a Comparable Increase to Bias Across Countries in the Programme for the International Assessment of Adult Competencies." JSM Proceedings, Survey Research Methods Section. Alexandria, VA: American Statistical Association (2014): 655-666. [Link](http://www.asasrms.org/Proceedings/y2014/files/311170_87007.pdf)

df_deso = pd.read_sql("SELECT deso, befolkning FROM public.zones", con=engine)
df_deso = df_deso.rename(columns={'befolkning': 'pop'})
df_deso_m = df_home_deso.groupby('deso').size().reset_index(name='count')
df_deso_m = pd.merge(df_deso, df_deso_m, how='left')
df_deso_m.fillna(0, inplace=True)

df_deso_m.loc[:, 'wt_p'] = df_deso_m.loc[:, 'pop'] / df_deso_m.loc[:, 'count']
df_deso_m.head()

### 3.1 Weight trimming

w0 = ((np.std(df_deso_m.loc[:, 'wt_p']) / np.mean(df_deso_m.loc[:, 'wt_p'])) ** 2 + 1) ** 0.5 * 3.5 * np.median(df_deso_m.loc[:, 'wt_p'])
df_deso_m.loc[df_deso_m['wt_p'] > w0, 'wt_p'] = w0

### 3.2 Get weight

df_home_deso = pd.merge(df_home_deso, df_deso_m[['deso', 'wt_p']], on='deso', how='left')

Save the data.

df_home_deso.drop(columns=['loc', 'holiday_s']).to_sql('home_p', engine, schema='public', index=False, method='multi', if_exists='replace', chunksize=10000)

## 5. Statistics on devices and actual population

df_home_deso = pd.read_sql("SELECT uid, deso FROM public.home_p", con=engine)
df_deso = pd.read_sql("SELECT deso, befolkning FROM public.zones", con=engine)
df_deso = df_deso.rename(columns={'befolkning': 'pop'})
df_deso_m = df_home_deso.groupby('deso').size().reset_index(name='count')
df_deso_m = pd.merge(df_deso, df_deso_m, how='left')
df_deso_m.fillna(0, inplace=True)
df_deso_m

from scipy import stats
res = stats.spearmanr(df_deso_m['pop'], df_deso_m['count'])
res

df_deso_m.loc[:, 'share'] = df_deso_m.loc[:, 'count'] / df_deso_m.loc[:, 'pop']
df_deso_m.describe()

df_deso_m.loc[df_deso_m['share'] < 1, 'share'].describe()
