==> src/MobiSegInsightsSE.etl.1.0/32-homophily-simulation-hex.ipynb <==
# Homophily and distance simulations at mixed-hexagon level
1) Randomly shift non-home activities within the same distance range;
2) Randomly shift non-home activiti with the average distance decay.

%load_ext autoreload
%autoreload 2
%cd D:\mobi-social-segregation-se

# Load libs
import pandas as pd
import os
os.environ['USE_PYGEOS'] = '0'
import preprocess
import geopandas as gpd
from tqdm.notebook import tqdm
import sqlalchemy
from collections import Counter
import random
import ast
from p_tqdm import p_map
import numpy as np
from scipy.spatial import distance
from sklearn.neighbors import KDTree
import seaborn as sns

# Data location
user = preprocess.keys_manager['database']['user']
password = preprocess.keys_manager['database']['password']
port = preprocess.keys_manager['database']['port']
db_name = preprocess.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}?gssencmode=disable')
# Data location for OSM data of Sweden (Aug 28, 2023)
db_name_osm = preprocess.keys_manager['osmdb']['name']
engine_osm = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name_osm}?gssencmode=disable')

## 1. Load data

# POIs in Sweden
gdf_pois = gpd.GeoDataFrame.from_postgis(sql="""SELECT osm_id, "Tag", geom FROM built_env.pois;""", con=engine)
gdf_pois = gdf_pois.to_crs(3006)
gdf_pois.loc[:, 'y'] = gdf_pois.geom.y
gdf_pois.loc[:, 'x'] = gdf_pois.geom.x

# Find which mixed-hexagon zone each POI belongs to
gdf_hex= gpd.GeoDataFrame.from_postgis(sql="""SELECT hex_id AS hex_s, deso, geom FROM spatial_units;""", con=engine)
gdf_hex = gdf_hex.to_crs(3006)
gdf_pois = gpd.sjoin(gdf_pois, gdf_hex[['hex_s', 'geom']])
gdf_pois.head()

# Load stops and add home label
df_stops = pd.read_sql(sql=f"""SELECT uid, lat, lng, wt_total, time_span, deso
                               FROM segregation.mobi_seg_deso_raw
                               WHERE weekday=1 AND holiday=0;""",
                       con=engine)
gdf_stops = preprocess.df2gdf_point(df_stops, 'lng', 'lat', crs=4326, drop=False)
df_home = pd.read_sql(sql=f"""SELECT uid, lat, lng, wt_p
                               FROM home_p;""",
                       con=engine)
df_home.loc[:, 'home'] = 1
gdf_stops = pd.merge(gdf_stops, df_home, on=['uid', 'lat', 'lng'], how='left')
gdf_stops = gdf_stops.fillna(0)

len(gdf_stops.loc[gdf_stops.wt_p==0, :]) / len(gdf_stops)

# Find hexagons for stops
deso_list = gdf_hex.loc[gdf_hex['hex_s'] == '0', 'deso'].values
hex_deso_list = gdf_hex.loc[gdf_hex['hex_s'] != '0', 'deso'].unique()
geo_deso = gdf_stops.loc[gdf_stops['deso'].isin(deso_list), :]
geo_hex = gdf_stops.loc[gdf_stops['deso'].isin(hex_deso_list), :]

geo_deso.loc[:, 'hex'] = geo_deso.loc[:, 'deso']

def find_hex(data):
    # uid, lat, lng, wt_total, dur, hex_id*
    deso = data.deso.values[0]
    gdf_d = data.copy()
    gdf_d = gpd.sjoin(gdf_d,
                      gdf_hex.loc[gdf_hex['deso'] == deso, :].drop(columns=['deso']).to_crs(4326),
                      how='inner')
    return gdf_d.rename(columns={'hex_s': 'hex'})

print("Find hexagons for geolocations by DeSO zone.")
tqdm.pandas()
geo4graph = geo_hex.groupby('deso').progress_apply(find_hex).reset_index(drop=True)
gdf_stops = pd.concat([geo4graph, geo_deso])
gdf_stops.head()

# Process stops
gdf_stops = gdf_stops.to_crs(3006)
gdf_stops.loc[:, 'y'] = gdf_stops.geometry.y
gdf_stops.loc[:, 'x'] = gdf_stops.geometry.x
print(len(gdf_stops))
gdf_stops.replace([np.inf, -np.inf], np.nan, inplace=True)
gdf_stops.dropna(subset=["x", "y"], how="any", inplace=True)
print("After processing infinite values", len(gdf_stops))

## 2. Find POI for each stop

gdf_pois = gdf_pois.reset_index(drop=True)
tree = KDTree(gdf_pois[["y", "x"]], metric="euclidean")
ind, dist = tree.query_radius(gdf_stops[["y", "x"]].to_records(index=False).tolist(),
                              r=300, return_distance=True, count_only=False, sort_results=True)
gdf_stops.loc[:, 'poi_num'] = [len(x) for x in ind]
gdf_stops.loc[gdf_stops.poi_num > 0, 'osm_id'] = [gdf_pois.loc[x[0], 'osm_id'] for x in ind if len(x) > 0]
gdf_stops.loc[gdf_stops.poi_num > 0, 'dist'] = [x[0] for x in dist if len(x) > 0]
gdf_stops = pd.merge(gdf_stops, gdf_pois[['osm_id', 'Tag']], on='osm_id', how='left')
gdf_stops.head()

print("Share of non-home stops with a nearby POI:")
len(gdf_stops.loc[(gdf_stops.home == 0) & (~gdf_stops.Tag.isna()), :]) / \
len(gdf_stops.loc[gdf_stops.home == 0, :])

## 3. Descriptive analysis on homophily theory

def group_ice_r(x):
    if x > 0.2:
        return 'D'
    elif x < -0.2:
        return 'F'
    else:
        return 'N'

df = pd.read_parquet('results/data4model_individual.parquet')
df = df.loc[(df.weekday == 1) & (df.holiday == 0), ['uid', 'wt_p', 'ice_birth_resi']].rename(columns={'ice_birth_resi': 'ice_r'})
df.loc[:, 'grp_r'] = df['ice_r'].apply(lambda x: group_ice_r(x))
df.iloc[0]

uids_seg = df.loc[df.grp_r != 'N', 'uid'].unique()
len(uids_seg)

### 3.1 Stay home share

def span2seq(time_seq_list):
    seq = list(range(time_seq_list[0], time_seq_list[1] + 1))
    if len(time_seq_list) > 2:
        seq2 = list(range(time_seq_list[2], time_seq_list[3] + 1))
        seq = seq2 + seq
    return seq

def home_share(data):
    df = data.copy()
    df.loc[:, 'time_span'] = df.loc[:, 'time_span'].apply(lambda x: ast.literal_eval(
        x.replace("{", "(").replace("}", ")")
    ))
    df.loc[:, 'time_seq'] = df.loc[:, 'time_span'].apply(span2seq)
    df.loc[:, 'dur'] = df.loc[:, 'time_seq'].apply(lambda x: len(x))
    df.loc[:, 'dur_wt'] = df.loc[:, 'dur'] * df.loc[:, 'wt_total']
    l_h = df.loc[df.home == 1, 'dur_wt'].sum()
    l_nh = df.loc[df.home == 0, 'dur_wt'].sum()
    return pd.Series(dict(home_share=l_h / (l_h + l_nh) * 100))

tqdm.pandas()
df_home_s = gdf_stops.loc[gdf_stops.uid.isin(uids_seg), ['uid', 'home', 'time_span', 'wt_total']].\
    groupby('uid').progress_apply(home_share).reset_index()

df = pd.merge(df, df_home_s, on='uid')
df.groupby('grp_r').apply(lambda x: pd.Series(dict(home_share=np.average(x['home_share'], weights=x['wt_p'])))).reset_index()

### 3.2 POI categories

def poi_share(data):
    return pd.Series(dict(dur=data.loc[:, 'dur_wt'].sum()))

def poi_share_grp(data):
    L = sum(data.dur * data.wt_p)
    return pd.Series(dict(tag_time=L))

def poi_share_marginal(data):
    L = data.tag_time.sum()
    data.loc[:, 'tag_time'] /= (L / 100)
    return data

df_poi = pd.merge(gdf_stops.loc[(gdf_stops.uid.isin(uids_seg)) & (gdf_stops.home == 0), 
                                ['uid', 'time_span', 'wt_total', 'Tag']], 
                  df, on='uid', how='left')
df_poi.loc[:, 'time_span'] = df_poi.loc[:, 'time_span'].apply(lambda x: ast.literal_eval(
    x.replace("{", "(").replace("}", ")")
))
df_poi.loc[:, 'time_seq'] = df_poi.loc[:, 'time_span'].apply(span2seq)
df_poi.loc[:, 'dur'] = df_poi.loc[:, 'time_seq'].apply(lambda x: len(x))
df_poi.loc[:, 'dur_wt'] = df_poi.loc[:, 'dur'] * df_poi.loc[:, 'wt_total']
df_poi.drop(columns=['time_span', 'time_seq', 'dur'], inplace=True)

df_poi_stats = df_poi.groupby(['uid', 'Tag']).progress_apply(poi_share).reset_index()


df_poi_stats = pd.merge(df_poi_stats, df, on='uid', how='left')
df_poi_stats.head()

df_poi_stats = df_poi_stats.groupby(['Tag', 'grp_r']).progress_apply(poi_share_grp).reset_index()

df_poi_stats = df_poi_stats.groupby('grp_r').progress_apply(poi_share_marginal).reset_index(drop=True)

df_poi_stats = df_poi_stats.sort_values(by=['tag_time', 'Tag', 'grp_r'], ascending=False)
df_poi_stats

df_poi_disp = pd.pivot_table(data=df_poi_stats, values=['tag_time'], index=['Tag'], columns=['grp_r']).reset_index()
df_poi_disp.columns = df_poi_disp.columns.droplevel()
df_poi_disp.loc[:, 'delta'] = df_poi_disp.loc[:, 'D'] - df_poi_disp.loc[:, 'F']
df_poi_disp = df_poi_disp.sort_values(by='delta', ascending=False)

df_poi_disp.loc[abs(df_poi_disp.delta) > 1, :]

## 4. Randomly shift non-home stops
To a similar POI within 1 km radius.

Step 1. If there are any POI having the same Tag, select one.
Step 2. If Step 1 fails, select one from any equivalent tags in amenity/shop class.
Step 3. If non in Office or Craft, select one from any tags in the other group.
Step 4. If they both below to shops, i.e., (s) or Shop.
Step 5. If all steps fail, set None for the stop.

### 4.1 POI randomization within distance range (1)

shift_radius = 1000 # m
shift_radius_lower = 30 # m

def poi2nearby(row):
    # Distance-constrained POI shifting
    X = gdf_pois.loc[gdf_pois.osm_id==row['osm_id'], 'x'].values[0]
    Y = gdf_pois.loc[gdf_pois.osm_id==row['osm_id'], 'y'].values[0]
    ind, dist = tree.query_radius([(Y, X)], r=shift_radius,
                                  return_distance=True, count_only=False, sort_results=True)
    ind = ind[0]
    dist = dist[0]
    def tag_categorization(x):
        if x == row['Tag']:
            return 1
        if ('(' in row['Tag']) & ('(' in x):
            if row['Tag'].split(' (')[0] == x.split(' (')[0]:
                return 2
        if (row['Tag'] in ('Office', 'Craft')) & (x in ('Office', 'Craft')):
            return 3
        if ('(s)' in row['Tag']) | (row['Tag'] == 'Shop'):
            if ('(s)' in x) | (x == 'Shop'):
                return 4
        return 0

    if len(ind) > 0:
        df = pd.DataFrame()
        df.loc[:, "id"] = range(0, len(ind))
        df.loc[:, "tag"] = [gdf_pois.loc[x, 'Tag'] for x in ind]
        df.loc[:, "dist"] = dist
        df.loc[:, "hex_s"] = [gdf_pois.loc[x, 'hex_s'] for x in ind]
        # Exclude POIs too close
        df = df.loc[df.dist > shift_radius_lower, :]
        if len(df) > 0:
            df.loc[:, "tag_cat"] = df.loc[:, "tag"].apply(lambda x: tag_categorization(x))
            if df.loc[:, "tag_cat"].sum() > 0:
                # sample 1 POI following the conditions 1, 2, 3, and 4
                for cat in (1, 2, 3, 4):
                    df2sample = df.loc[df.tag_cat == cat, :]
                    if len(df2sample) > 0:
                        hex_pool = random.choices(df2sample['hex_s'].values, k=100)
                        # hex_st = df2sample['hex_s'].values[0]
                        break
            else:
                hex_pool = []
                # hex_st = ''
        else:
            hex_pool = []
            # hex_st = ''
            
        if len(hex_pool) > 0:
            hx_str = str(dict(Counter(hex_pool)))
        else:
            hx_str = ''
    
        return pd.Series(dict(hex_s=hx_str))

stops2shift = gdf_stops.loc[(gdf_stops.home == 0) & (~gdf_stops.Tag.isna()), :]
stops2keep = gdf_stops.loc[~((gdf_stops.home == 0) & (~gdf_stops.Tag.isna())), :]

#### 4.1.2 Run the homophily simulation for the entire dataset and save

tqdm.pandas()
shifted = stops2shift.sample(1000).progress_apply(poi2nearby, axis=1)
shifted.head()

stops2shift = pd.concat([stops2shift, shifted], axis=1)

gdf_stops = pd.concat([stops2shift, stops2keep])
gdf_stops.iloc[0]

gdf_stops.drop(columns=['x', 'y', 'geometry']).\
    to_sql('mobi_seg_hex_raw_sim1_w1h0', engine, schema='segregation', index=False,
           method='multi', if_exists='append', chunksize=10000)

### 3.2 Distance-decay randomization (2)
MobiSegIndightsSE.etl.1.0/`32-homophily-simulation-dist-free.py`
