==> src/MobiSegInsightsSE.etl.1.0/32-pois-preference-shifted.ipynb <==
# Extract POIs for the random move model
The data are from OpenStreetMap, loaded to postgres, using osm2pgsql-1.9.1-x64.

`osm2pgsql -d osm_sweden -U postgres -W -H localhost -P 5433 -S D:\mobi-social-segregation-se\src\MobiSegInsightsSE.etl.1.0\osm2pgsql-1.9.1-x64\osm2pgsql-bin\default.style D:\mobi-social-segregation-se\dbs\geo\sweden-latest.osm.pbf`

The below command line extracts POIs from OSM data in Sweden.
`osm2pgsql -d osm_sweden -U postgres -W -H localhost -P 5433 -S D:\mobi-social-segregation-se\src\MobiSegInsightsSE.etl.1.0\osm2pgsql-1.9.1-x64\osm2pgsql-bin\flex-config\pois.lua -O flex D:\mobi-social-segregation-se\dbs\geo\sweden-latest.osm.pbf`


%load_ext autoreload
%autoreload 2
%cd D:\mobi-social-segregation-se

# Load libs
import pandas as pd
import os
os.environ['USE_PYGEOS'] = '0'
import openai
import preprocess
import geopandas as gpd
from tqdm.notebook import tqdm
import sqlalchemy
import time
import numpy as np
from sklearn.neighbors import KDTree

openai.api_key = preprocess.keys_manager['openai']['key']

# Data location
user = preprocess.keys_manager['database']['user']
password = preprocess.keys_manager['database']['password']
port = preprocess.keys_manager['database']['port']
db_name = preprocess.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}?gssencmode=disable')

# Data location for OSM data of Sweden (Aug 28, 2023)
db_name_osm = preprocess.keys_manager['osmdb']['name']
engine_osm = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name_osm}?gssencmode=disable')

## 1. POIs

# Get pois from database
gdf_pois = gpd.GeoDataFrame.from_postgis(sql="""SELECT osm_id, "class", subclass, geom FROM pois;""", con=engine_osm)
gdf_pois.head()

df_pois_tp = gdf_pois.groupby(['class', 'subclass']).size().to_frame(name='count').reset_index()

len(df_pois_tp)

df_pois_tp

tags = [x for x in df_pois_tp.subclass.values]
tags[4]

### 1.1 Define broader categories
This is learned from all the tags by ChatGPT.

categories_str = "- Healthcare\n- Financial Services\n- Food and Drink\n- Entertainment\n- Education\n- Recreation\n- Transportation\n- Religious Places\n- Emergency Services\n- Artisan Workshops\n- Automotive Services\n- Craft\n- Historic\n- Sports and Activities\n- Outdoor Recreation\n- Leisure\n- Office\n- Groceries and Food\n- Fashion and Accessories\n- Home and Living\n- Health and Beauty\n- Indoor Sports\n- Sports\n- Tourism"
categories = categories_str.split('\n')
len(categories)

### 1.2 Relabel POIs with the new categories based on subclass
This is done by GPT-4.

def poi_category(x):
    flag = 0
    while flag != 1:
        try:
            response = openai.ChatCompletion.create(
              model="gpt-4",
              messages=[
                {
                  "role": "system",
                  "content": f"You will be presented with points of interest tags from OpenStreetMap and your job is to provide the most suitable tag from the following list. Choose ONLY from the list of tags provided here:\n\n{categories_str}"
                },
                {
                  "role": "user",
                  "content": x
                }
              ],
              temperature=0,
              max_tokens=1024,
              top_p=1,
              frequency_penalty=0,
              presence_penalty=0
            )
            flag = 1
            cate = response.choices[0].message.content
        except:
            time.sleep(1)
    return cate

tqdm.pandas()
df_pois_tp.loc[:, 'category'] = df_pois_tp.loc[:, 'subclass'].progress_apply(lambda x: poi_category(x))

df_pois_tp.to_csv('results/pois/tag_category.csv', index=False)

### 1.3 Manually check every re-labeled class-subclass
This step cleaned up some tags and decided which POIs to exclude.

Only those POIs with long duration of stay are included. For example, emergency points and transportation toll are removed.

df_pois_tp = pd.read_csv('results/pois/tag_category.csv')
df_pois_tp.head()

df_pois_tp = df_pois_tp.loc[df_pois_tp['Keep'] == 1, :]

df_pois_tp_r = df_pois_tp.groupby(['class', 'category'])['count'].sum().reset_index()
df_pois_tp_r.to_clipboard(index=False)

### 1.4 Merge some classes

Do not distinguish categories: tourism, historic = tourism, leisure, sport = leisure, craft, office = office.

Distinguish categories: amenity, shop.

df_pois_tp_red = pd.read_excel('results/pois/tags_reduced.xlsx', sheet_name='Sheet1')
df_pois_tp_red.head()

### 1.5 Add final tag to POI data

gdf_pois = pd.merge(gdf_pois, df_pois_tp[['class', 'subclass', 'category']], how='inner')
gdf_pois = pd.merge(gdf_pois, df_pois_tp_red[['class', 'category', 'Tag']], how='inner')
gdf_pois.head()

gdf_pois.to_crs(4326).to_postgis("pois", schema="built_env", con=engine)

gdf_pois.groupby('Tag').size()

## 2. Find nearest POIs
Search radius = 300 m

gdf_pois = gpd.GeoDataFrame.from_postgis(sql="""SELECT osm_id, "Tag", geom FROM built_env.pois;""", con=engine)
gdf_pois = gdf_pois.to_crs(3006)
gdf_pois.loc[:, 'y'] = gdf_pois.geom.y
gdf_pois.loc[:, 'x'] = gdf_pois.geom.x
gdf_pois.head()

### 2.1 Load stops

df_stops = pd.read_sql(sql=f"""SELECT uid, lat, lng, wt_total, time_span, deso
                               FROM segregation.mobi_seg_deso_raw
                               WHERE weekday=1 AND holiday=0;""",
                       con=engine)
gdf_stops = preprocess.df2gdf_point(df_stops, 'lng', 'lat', crs=4326, drop=False)

Add home label to the stops

df_home = pd.read_sql(sql=f"""SELECT uid, lat, lng
                               FROM home_p;""",
                       con=engine)
df_home.loc[:, 'home'] = 1
gdf_stops = pd.merge(gdf_stops, df_home, on=['uid', 'lat', 'lng'], how='left')
gdf_stops = gdf_stops.fillna(0)
gdf_stops.head()

gdf_stops.groupby('home').size()

### 2.2 Create a KD tree of all POIs and find the nearest POI within 300 m radius

gdf_stops = gdf_stops.to_crs(3006)
gdf_stops.loc[:, 'y'] = gdf_stops.geometry.y
gdf_stops.loc[:, 'x'] = gdf_stops.geometry.x

print(len(gdf_stops))
gdf_stops.replace([np.inf, -np.inf], np.nan, inplace=True)
gdf_stops.dropna(subset=["x", "y"], how="any", inplace=True)
print("After processing infinite values", len(gdf_stops))

tree = KDTree(gdf_pois[["y", "x"]], metric="euclidean")

ind, dist = tree.query_radius(gdf_stops[["y", "x"]].to_records(index=False).tolist(),
                              r=300, return_distance=True, count_only=False, sort_results=True)

gdf_stops.loc[:, 'poi_num'] = [len(x) for x in ind]
gdf_stops.loc[gdf_stops.poi_num > 0, 'osm_id'] = [gdf_pois.loc[x[0], 'osm_id'] for x in ind if len(x) > 0]
gdf_stops.loc[gdf_stops.poi_num > 0, 'dist'] = [x[0] for x in dist if len(x) > 0]
gdf_stops = pd.merge(gdf_stops, gdf_pois[['osm_id', 'Tag']], on='osm_id', how='left')
gdf_stops.head()

print("Share of non-home stops with a nearby POI:")
len(gdf_stops.loc[(gdf_stops.home == 0) & (~gdf_stops.Tag.isna()), :]) / \
len(gdf_stops.loc[gdf_stops.home == 0, :])

## 3. Randomly shift non-home stops
To a similar POI within 1 km radius.

Step 1. If there are any POI having the same Tag, select one.
Step 2. If Step 1 fails, select one from any equivalent tags in amenity/shop class.
Step 3. If non in Office or Craft, select one from any tags in the other group.
Step 4. If they both below to shops, i.e., (s) or Shop.
Step 5. If all steps fail, set None for the stop.

shift_radius = 1000 # m
shift_radius_lower = 30 # m

def poi2nearby(row):
    X = gdf_pois.loc[gdf_pois.osm_id==row['osm_id'], 'x'].values[0]
    Y = gdf_pois.loc[gdf_pois.osm_id==row['osm_id'], 'y'].values[0]
    ind, dist = tree.query_radius([(Y, X)], r=shift_radius,
                                  return_distance=True, count_only=False, sort_results=True)
    ind = ind[0]
    dist = dist[0]
    def tag_categorization(x):
        if x == row['Tag']:
            return 1
        if ('(' in row['Tag']) & ('(' in x):
            if row['Tag'].split(' (')[0] == x.split(' (')[0]:
                return 2
        if (row['Tag'] in ('Office', 'Craft')) & (x in ('Office', 'Craft')):
            return 3
        if ('(s)' in row['Tag']) | (row['Tag'] == 'Shop'):
            if ('(s)' in x) | (x == 'Shop'):
                return 4
        return 0

    if len(ind) > 0:
        df = pd.DataFrame()
        df.loc[:, "id"] = range(0, len(dist))
        df.loc[:, "dist"] = dist
        df.loc[:, "tag"] = [gdf_pois.loc[x, 'Tag'] for x in ind]
        df.loc[:, "osm_id"] = [gdf_pois.loc[x, 'osm_id'] for x in ind]
        # Exclude POIs too close
        df = df.loc[df.dist > shift_radius_lower, :]
        if len(df) > 0:
            df.loc[:, "tag_cat"] = df.loc[:, "tag"].apply(lambda x: tag_categorization(x))
            if df.loc[:, "tag_cat"].sum() > 0:
                # sample 1 POI following the conditions 1, 2, 3, and 4
                for cat in (1, 2, 3, 4):
                    df2sample = df.loc[df.tag_cat == cat, :]
                    if len(df2sample) > 0:
                        poi_pool = len(df2sample)
                        df2sample = df2sample.sample(1)
                        osm_id_s = df2sample.osm_id.values[0]
                        Tag_s = df2sample.tag.values[0]
                        dist2poi = df2sample.dist.values[0]
                        break
            else:
                osm_id_s = np.nan
                Tag_s = np.nan
                dist2poi = np.nan
                poi_pool = 0
        else:
            osm_id_s = np.nan
            Tag_s = np.nan
            dist2poi = np.nan
            poi_pool = 0
        return pd.Series([osm_id_s, Tag_s, dist2poi, poi_pool], index=['osm_id_s', 'Tag_s', 'dist2poi', 'poi_pool'])

stops2shift = gdf_stops.loc[(gdf_stops.home == 0) & (~gdf_stops.Tag.isna()), :]
stops2keep = gdf_stops.loc[~((gdf_stops.home == 0) & (~gdf_stops.Tag.isna())), :]

test = False
if test:
    stops2shift = stops2shift.sample(1000)

tqdm.pandas()
shifted = stops2shift.progress_apply(poi2nearby, axis=1)
stops2shift = pd.concat([stops2shift, shifted], axis=1)

gdf_stops = pd.concat([stops2shift, stops2keep])
gdf_stops.iloc[0]

### 3.1 Find the corresponding DeSO zones for the shifted stays

gdf_deso = gpd.GeoDataFrame.from_postgis(sql="""SELECT deso AS deso_s, geom FROM zones;""", con=engine)
gdf_pois = gpd.sjoin(gdf_pois, gdf_deso)
gdf_pois.head()

gdf_stops = pd.merge(gdf_stops, gdf_pois[['osm_id', 'deso_s']].rename(columns={'osm_id': 'osm_id_s'}),
                     on='osm_id_s', how='left')
gdf_stops.iloc[0]

print("Share of non-home stops with a shifted POI:")
len(gdf_stops.loc[(gdf_stops.home == 0) & (gdf_stops.poi_pool > 0), :]) / \
len(gdf_stops.loc[gdf_stops.home == 0, :])

print("Share of non-home stops with a shifted POI in different DeSO zone:")
len(gdf_stops.loc[(gdf_stops.home == 0) & (gdf_stops.poi_pool > 0) & (gdf_stops.deso_s != gdf_stops.deso), :]) / \
len(gdf_stops.loc[(gdf_stops.home == 0) & (gdf_stops.poi_pool > 0), :])

gdf_stops.drop(columns=['x', 'y', 'geometry']).\
    to_sql('mobi_seg_deso_raw_poi_w1h0', engine, schema='segregation', index=False,
           method='multi', if_exists='append', chunksize=10000)

print("Number of individuals that have at least one shifted stay on a different DeSO zone.")
gdf_stops.loc[(gdf_stops.home == 0) & (gdf_stops.poi_pool > 0) & (gdf_stops.deso_s != gdf_stops.deso), 'uid'].nunique()
